chore(users): move enum class to graphql file
